Move non-genome probes counts and metadata to altExp slots

moveNonGenomeProbes() takes the probe IDs corresponding to grnaProbe and barcodeProbe slots of the TapestriExperiment object, as well as probes on chrY, and moves them to their own altExp slots in the object. This allows those counts and associated metadata to be manipulated separately without interfering with the probes used for CNV measurements which target the endogenous genome. SingleCellExperiment::splitAltExps() can be used for manual specification of probes to move to altExp slots if the shortcut slots are not used.

Usage

moveNonGenomeProbes(TapestriExperiment)

Arguments

TapestriExperiment

TapestriExperiment object.

Value

TapestriExperiment with altExp slots filled with counts and metadata for non-genomic probes.

Details

moveNonGenomeProbes() moves probes corresponding to the specified tags to altExp (alternative experiment) slots in the TapestriExperiment object. These probes should be those which do not correspond to a chromosome and therefore would not be used to call copy number variants. The exception is probes on chromosome Y; CNVs of chrY are more rare, so we move it to an altExp for separate analysis. Probes corresponding to the barcodeProbe and grnaProbe slots, which are specified by the panel.id shortcut or manually (see Custom Slot Getters and Setters), are automatically moved to altExp by this operation as well. If such probes are not present, the function will only generate a warning message, so it is always safe (and recommended) to run by default. Any remaining probes that are not targeting a human chromosome and are not specified by the shortcut tags are moved to the otherProbeCounts slot. This function is run automatically by default and with default behavior as part of createTapestriExperiment().

See also

Examples

tap.object <- newTapestriExperimentExample() # example TapestriExperiment
#> ℹ Moving gRNA probe  to `altExp` slot "grnaCounts".
#> ℹ Moving barcode probe  to `altExp` slot "barcodeCounts".
#> ℹ Moving chrY probe(s) probe_231, probe_232, probe_233, probe_234, probe_235, probe_236, probe_237, probe_238, probe_239, and probe_240 to `altExp` slot "chrYCounts".
tap.object <- moveNonGenomeProbes(tap.object)
#> ℹ Moving gRNA probe  to `altExp` slot "grnaCounts".
#> ℹ Moving barcode probe  to `altExp` slot "barcodeCounts".
#> ℹ ChrY probe ID(s) not found in TapestriExperiment object.
#> ℹ No non-genomic probe IDs found.