Getter and Setter functions for TapestriExperiment slots

Get and set custom slots in TapestriExperiment. Slots include barcodeProbe for a sample barcode probe ID and grnaProbe for a gRNA-associated probe ID. These are used as shortcuts for moveNonGenomeProbes() and countBarcodedReads(). gmmParams holds parameters and metadata for GMM copy number calling models.

Usage

barcodeProbe(x)

# S4 method for class 'TapestriExperiment'
barcodeProbe(x)

barcodeProbe(x) <- value

# S4 method for class 'TapestriExperiment'
barcodeProbe(x) <- value

grnaProbe(x)

# S4 method for class 'TapestriExperiment'
grnaProbe(x)

grnaProbe(x) <- value

# S4 method for class 'TapestriExperiment'
grnaProbe(x) <- value

gmmParams(x)

# S4 method for class 'TapestriExperiment'
gmmParams(x)

Arguments

x

A TapestriExperiment object

value

Character, probe ID to assign to slot

TapestriExperiment

A TapestriExperiment object

Value

For the getter methods barcodeProbe, grnaProbe, and gmmParams, the value of the given slot is returned. For the setter methods barcodeProbe and grnaProbe, a TapestriExperiment object is returned with modifications made to the given slot.

Functions

barcodeProbe(TapestriExperiment): barcodeProbe getter

barcodeProbe(TapestriExperiment) <- value: barcodeProbe setter

grnaProbe(TapestriExperiment): grnaProbe getter

grnaProbe(TapestriExperiment) <- value: grnaProbe setter

gmmParams(TapestriExperiment): gmmParams getter

Examples

tap.object <- newTapestriExperimentExample() # example TapestriExperiment object
#> ℹ Moving gRNA probe  to `altExp` slot "grnaCounts".
#> ℹ Moving barcode probe  to `altExp` slot "barcodeCounts".
#> ℹ Moving chrY probe(s) probe_231, probe_232, probe_233, probe_234, probe_235, probe_236, probe_237, probe_238, probe_239, and probe_240 to `altExp` slot "chrYCounts".
barcodeProbe(tap.object) <- "Probe01"
barcodeProbe(tap.object)
#> [1] "Probe01"

grnaProbe(tap.object) <- "Probe02"
grnaProbe(tap.object)
#> [1] "Probe02"

gmmParams(tap.object)
#> list()